[Team-5 BE 익조 & 피오] Sidedish 프로젝트 1주차(금요일) 코드 리뷰 요청

+++ BE/src/main/java/com/sidedish/dto/EventCategoryDto.java
@@ -5,10 +5,10 @@ import lombok.Getter;
 @Getter
 public class EventCategoryDto {
 
-    private Long id;
+    private Integer id;
     private String name;
 
-    public EventCategoryDto(Long eventCategoryId, String eventCategoryName) {
+    public EventCategoryDto(Integer eventCategoryId, String eventCategoryName) {
         this.id = eventCategoryId;
         this.name = eventCategoryName;
     }
